Rework plugin goals
Maven deletes the scope=import entries from the dependencyManagement
section of the project model. Also, new entries added by the import-bom
goal are not processed further on, so there is no point modifying that
list from the plugin.

For the reason above we decided to drop the import-bom goal, and only
leave the import-properties mojo.

For the same reason the new import-bom-properties goal cannot be
properly implemented, since it cannot scan the original dependency
management list from the current project and recursively.

For now users will have to duplicate the list of BOM imports and copy
the same list within the artifacts config entry for this plugin.